Refactor: move T4 C# code generation result
