+++ Backend/ForTea.Core/TemplateProcessing/CodeCollecting/T4CSharpCodeGenerationInfoCollector.cs
@@ -1,5 +1,6 @@
 using GammaJul.ForTea.Core.Psi;
 using GammaJul.ForTea.Core.Psi.Directives;
+using GammaJul.ForTea.Core.TemplateProcessing.CodeGeneration;
 using GammaJul.ForTea.Core.Tree;
 using JetBrains.Annotations;
