+++ Backend/ForTea.Core/TemplateProcessing/CodeCollecting/T4CSharpCodeGenerationInfoCollectorBase.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using GammaJul.ForTea.Core.Psi;
 using GammaJul.ForTea.Core.Psi.Directives;
+using GammaJul.ForTea.Core.TemplateProcessing.CodeGeneration;
 using GammaJul.ForTea.Core.Tree;
 using GammaJul.ForTea.Core.Tree.Impl;
 using JetBrains.Annotations;
